+++ govtool/backend/src/VVA/API.hs
@@ -5,9 +5,12 @@
 {-# LANGUAGE RecordWildCards   #-}
 {-# LANGUAGE TypeOperators     #-}
 {-# LANGUAGE ViewPatterns      #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 
 module VVA.API where
 
+import qualified Network.WebSockets.Connection as WS
+import Servant.API.WebSocket (WebSocket)
 import Control.Concurrent.QSem (waitQSem, signalQSem)
 import Control.Concurrent.Async (mapConcurrently)
 import           Control.Exception    (throw, throwIO)
@@ -44,6 +47,7 @@ import           VVA.Types            (App, AppEnv (..),
                                        AppError (CriticalError, ValidationError, InternalError),
                                        CacheEnv (..))
 import qualified VVA.Metadata         as Metadata
+import Servant.OpenApi (HasOpenApi, toOpenApi)
 
 type VVAApi =
          "drep" :> "list"
@@ -78,6 +82,8 @@ type VVAApi =
     :<|> "network" :> "metrics" :> Get '[JSON] GetNetworkMetricsResponse
     :<|> "proposal" :> "metadata" :> "validate" :> ReqBody '[JSON] MetadataValidationParams :> Post '[JSON] MetadataValidationResponse
     :<|> "drep" :> "metadata" :> "validate" :> ReqBody '[JSON] MetadataValidationParams :> Post '[JSON] MetadataValidationResponse
+    :<|> "transaction" :> "watch" :> Capture "transactionId" HexText :> WebSocket
+
 server :: App m => ServerT VVAApi m
 server = drepList
     :<|> getVotingPower
@@ -93,6 +99,20 @@ server = drepList
     :<|> getNetworkMetrics
     :<|> getProposalMetadataValidationResponse
     :<|> getDRepMetadataValidationResponse
+    :<|> transactionWatch
+
+instance HasOpenApi WebSocket where
+  toOpenApi _ = mempty
+
+
+transactionWatch :: App m => HexText -> WS.Connection -> m ()
+transactionWatch (unHexText -> transactionId) c = do
+  tvar <- asks vvaWebSocketConnections
+  Transaction.watchTransaction tvar transactionId c
+  liftIO $ forever $ do
+        msg <- WS.receiveData c
+        putStrLn $ Text.unpack $ ("Received: " <> msg)
+        WS.sendTextData c (msg :: Text)
 
 
 mapDRepType :: Types.DRepType -> DRepType